Se añade lógica al entorno de DeleteStudentUseCase.kt

+++ app/src/main/java/edu/iesam/studentplayground/features/students/domain/DeleteStudentUseCase.kt
@@ -1,7 +1,15 @@
 package edu.iesam.studentplayground.features.students.domain
 
+import android.util.Log
+
 class DeleteStudentUseCase(val studentRepository: StudentRepository) {
     operator fun invoke(exp:String){
-        studentRepository.delete(exp)
+        val existingStudent: Student? = studentRepository.getStudent(exp)
+
+        if(existingStudent != null) {
+            studentRepository.delete(exp)
+        }else{
+            Log.d("@dev", "El usuario que se intenta eliminar no existe")
+        }
     }
 }
